core/requirements.txt
Django==4.2
djangorestframework
Pillow
pyjwt